+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add tests for matching prefix values
 * Change prefix matching to use regex (with anchor) instead of fixed value
+* Add tests for prefixes with wildcards
 
 ## clustree 0.4.1.9001 (2020-01-27)